refactor: 상품 상세 페이지에서 장바구니에 담는 로직 수정

+++ app/src/main/java/woowacourse/shopping/feature/goods/GoodsActivity.kt
@@ -147,9 +147,9 @@ class GoodsActivity : AppCompatActivity() {
     }
 
     private fun handleActivityResult(data: Intent?) {
-        val changedId = data?.getIntExtra("GOODS_ID", 0) ?: 0
+        val changedId = data?.getLongExtra("GOODS_ID", 0) ?: 0
         val changedQuantity = data?.getIntExtra("GOODS_QUANTITY", 0) ?: 0
-        viewModel.updateItemQuantity(changedId, changedQuantity)
+        viewModel.updateItemQuantity(changedId.toInt(), changedQuantity)
         viewModel.refreshHistoryOnly()
     }
